OpenaiEvaluation/requirements.txt
httpx==0.27.2
playwright==1.46.0
openai==1.40.6
pydantic==2.8.2
redis==5.0.7
zstandard==0.23.0
google-cloud-storage==2.18.2